agent/requirements.txt
python-dotenv>=1.0.0
gql[aiohttp]>=3.4.0
requests>=2.31.0